move everything to extension; add tests frule
