+++ Project.toml
@@ -5,8 +5,6 @@ version = "0.12.2"
 
 [deps]
 CEnum = "fa961155-64e5-5f13-b03f-caf6b980ea82"
-ChainRules = "082447d4-558c-5d27-93f4-14fc19e9eca2"
-ChainRulesCore = "d360d2e6-b24c-11e9-a2a3-2a2ae2dbcce4"
 EnzymeCore = "f151be2c-9106-41f4-ab19-57ee4f262869"
 Enzyme_jll = "7cc45869-7501-5eee-bdea-0790c847d4ef"
 GPUCompiler = "61eb1bfa-7361-4325-ad38-22787b887f55"
@@ -19,13 +17,16 @@ Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 
 [weakdeps]
+ChainRulesCore = "d360d2e6-b24c-11e9-a2a3-2a2ae2dbcce4"
 SpecialFunctions = "276daf66-3868-5448-9aa4-cd146d93841b"
 
 [extensions]
 EnzymeSpecialFunctionsExt = "SpecialFunctions"
+EnzymeChainRulesCoreExt = "ChainRulesCore"
 
 [compat]
 CEnum = "0.4, 0.5"
+ChainRulesCore = "1"
 EnzymeCore = "0.7"
 Enzyme_jll = "0.0.105"
 GPUCompiler = "0.21, 0.22, 0.23, 0.24, 0.25, 0.26"
