+++ test/ext/specialfunctions.jl
